Check for Watcher; move function to upgrade package
