+++ internal/pkg/agent/application/upgrade/upgrade_test.go
@@ -5,6 +5,7 @@
 package upgrade
 
 import (
+	"context"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -13,9 +14,13 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/elastic/elastic-agent/pkg/control/v2/client"
+	"github.com/elastic/elastic-agent/pkg/control/v2/cproto"
+
 	"github.com/gofrs/flock"
 	"github.com/stretchr/testify/require"
 
+	"github.com/elastic/elastic-agent/internal/pkg/agent/errors"
 	"github.com/elastic/elastic-agent/internal/pkg/release"
 	"github.com/elastic/elastic-agent/pkg/core/logger"
 )
@@ -133,4 +138,95 @@ func TestShutdownCallback(t *testing.T) {
 	newContent, err := ioutil.ReadFile(newFilename)
 	require.NoError(t, err, "reading file failed")
 	require.Equal(t, content, newContent, "contents are not equal")
+}
+
+type mockClient struct {
+	stateErr string
+	state    cproto.State
+}
+
+func (mc *mockClient) Connect(ctx context.Context) error { return nil }
+func (mc *mockClient) Disconnect()                       {}
+func (mc *mockClient) Version(ctx context.Context) (client.Version, error) {
+	return client.Version{}, nil
+}
+func (mc *mockClient) State(ctx context.Context) (*client.AgentState, error) {
+	if mc.stateErr != "" {
+		return nil, errors.New(mc.stateErr)
+	}
+
+	return &client.AgentState{State: mc.state}, nil
+}
+func (mc *mockClient) StateWatch(ctx context.Context) (client.ClientStateWatch, error) {
+	return nil, nil
+}
+func (mc *mockClient) Restart(ctx context.Context) error { return nil }
+func (mc *mockClient) Upgrade(ctx context.Context, version string, sourceURI string, skipVerify bool, skipDefaultPgp bool, pgpBytes ...string) (string, error) {
+	return "", nil
+}
+func (mc *mockClient) DiagnosticAgent(ctx context.Context, additionalDiags []client.AdditionalMetrics) ([]client.DiagnosticFileResult, error) {
+	return nil, nil
+}
+func (mc *mockClient) DiagnosticUnits(ctx context.Context, units ...client.DiagnosticUnitRequest) ([]client.DiagnosticUnitResult, error) {
+	return nil, nil
+}
+func (mc *mockClient) DiagnosticComponents(ctx context.Context, additionalDiags []client.AdditionalMetrics, components ...client.DiagnosticComponentRequest) ([]client.DiagnosticComponentResult, error) {
+	return nil, nil
+}
+func (mc *mockClient) Configure(ctx context.Context, config string) error { return nil }
+
+func TestIsInProgress(t *testing.T) {
+	tests := map[string]struct {
+		state              cproto.State
+		stateErr           string
+		watcherPIDsFetcher func() ([]int, error)
+
+		expected    bool
+		expectedErr string
+	}{
+		"state_error": {
+			state:              cproto.State_STARTING,
+			stateErr:           "some error",
+			watcherPIDsFetcher: func() ([]int, error) { return nil, nil },
+
+			expected:    false,
+			expectedErr: "failed to get agent state: some error",
+		},
+		"state_upgrading": {
+			state:              cproto.State_UPGRADING,
+			stateErr:           "",
+			watcherPIDsFetcher: func() ([]int, error) { return nil, nil },
+
+			expected:    true,
+			expectedErr: "",
+		},
+		"state_healthy_no_watcher": {
+			state:              cproto.State_HEALTHY,
+			stateErr:           "",
+			watcherPIDsFetcher: func() ([]int, error) { return []int{}, nil },
+
+			expected:    false,
+			expectedErr: "",
+		},
+		"state_healthy_with_watcher": {
+			state:              cproto.State_HEALTHY,
+			stateErr:           "",
+			watcherPIDsFetcher: func() ([]int, error) { return []int{9999}, nil },
+
+			expected:    true,
+			expectedErr: "",
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			mc := mockClient{state: test.state, stateErr: test.stateErr}
+			inProgress, err := IsInProgress(&mc, test.watcherPIDsFetcher)
+			if test.expectedErr != "" {
+				require.Equal(t, test.expectedErr, err.Error())
+			} else {
+				require.Equal(t, test.expected, inProgress)
+			}
+		})
+	}
 }
